+++ libraries/matchers.rb
@@ -34,7 +34,7 @@ if defined?(ChefSpec)
     oneview_network_set:                standard_actions,
     oneview_rack:                       [:add, :remove, :add_if_missing, :add_to_rack, :remove_from_rack],
     oneview_server_hardware:            [:add_if_missing, :remove, :refresh, :set_power_state, :update_ilo_firmware],
-    oneview_storage_pool:               [:add, :remove],
+    oneview_storage_pool:               [:add_if_missing, :remove],
     oneview_storage_system:             [:add, :remove, :edit_credentials, :add_if_missing],
     oneview_switch:                     [:remove, :none],
     oneview_volume:                     standard_actions,