+++ spec/fixtures/cookbooks/oneview_test/recipes/storage_pool_add_if_missing_with_ip.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: oneview_test
-# Recipe:: storage_pool_add_with_name
+# Recipe:: storage_pool_add_with_ip
 #
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
@@ -14,7 +14,7 @@
 # specific language governing permissions and limitations under the License.
 #
 
-oneview_storage_pool 'StoragePool2' do
+oneview_storage_pool 'StoragePool1' do
   client node['oneview_test']['client']
-  storage_system_name 'StorageSystem1'
+  storage_system '10.1.1.1'
 end